+++ boa_engine/src/environments/runtime/declarative/lexical.rs
@@ -11,7 +11,7 @@ pub(crate) struct LexicalEnvironment {
 
 impl LexicalEnvironment {
     /// Creates a new `LexicalEnvironment`.
-    pub(crate) fn new(bindings: usize, poisoned: bool, with: bool) -> Self {
+    pub(crate) fn new(bindings: u32, poisoned: bool, with: bool) -> Self {
         Self {
             inner: PoisonableEnvironment::new(bindings, poisoned, with),
         }
@@ -28,7 +28,7 @@ impl LexicalEnvironment {
     ///
     /// Panics if the binding value is out of range or not initialized.
     #[track_caller]
-    pub(crate) fn get(&self, index: usize) -> Option<JsValue> {
+    pub(crate) fn get(&self, index: u32) -> Option<JsValue> {
         self.inner.get(index)
     }
 
@@ -38,7 +38,7 @@ impl LexicalEnvironment {
     ///
     /// Panics if the binding value is out of range.
     #[track_caller]
-    pub(crate) fn set(&self, index: usize, value: JsValue) {
+    pub(crate) fn set(&self, index: u32, value: JsValue) {
         self.inner.set(index, value);
     }
 }
