+++ boa_engine/src/object/internal_methods/arguments.rs
@@ -41,7 +41,7 @@ pub(crate) fn arguments_exotic_get_own_property(
             .borrow()
             .as_mapped_arguments()
             .expect("arguments exotic method must only be callable from arguments objects")
-            .get(*index as usize)
+            .get(*index)
         {
             // a. Set desc.[[Value]] to Get(map, P).
             return Ok(Some(
@@ -78,8 +78,8 @@ pub(crate) fn arguments_exotic_define_own_property(
         obj.borrow()
             .as_mapped_arguments()
             .expect("arguments exotic method must only be callable from arguments objects")
-            .get(index as usize)
-            .map(|value| (index as usize, value))
+            .get(index)
+            .map(|value| (index, value))
     } else {
         None
     };
@@ -170,7 +170,7 @@ pub(crate) fn arguments_exotic_get(
             .borrow()
             .as_mapped_arguments()
             .expect("arguments exotic method must only be callable from arguments objects")
-            .get(*index as usize)
+            .get(*index)
         {
             // a. Assert: map contains a formal parameter mapping for P.
             // b. Return Get(map, P).
@@ -209,7 +209,7 @@ pub(crate) fn arguments_exotic_set(
             obj.borrow_mut()
                 .as_mapped_arguments_mut()
                 .expect("arguments exotic method must only be callable from arguments objects")
-                .set(index as usize, &value);
+                .set(index, &value);
         }
     }
 
@@ -240,7 +240,7 @@ pub(crate) fn arguments_exotic_delete(
             obj.borrow_mut()
                 .as_mapped_arguments_mut()
                 .expect("arguments exotic method must only be callable from arguments objects")
-                .delete(*index as usize);
+                .delete(*index);
         }
     }
